+++ src/test/java/net/ttddyy/dsproxy/transform/CallableStatementQueryTransformerTest.java
@@ -103,7 +103,7 @@ public class CallableStatementQueryTransformerTest {
         QueryExecutionListener queryListener = mock(QueryExecutionListener.class);
         InterceptorHolder interceptorHolder = new InterceptorHolder(queryListener, transformer);
 
-        return new JdkJdbcProxyFactory().createConnection(rawDatasource.getConnection(), interceptorHolder);
+        return new JdkJdbcProxyFactory().createConnection(rawDatasource.getConnection(), interceptorHolder, "myDS");
     }
 
     @Test