+++ src/test/java/net/ttddyy/dsproxy/transform/PreparedStatementParameterTransformTest.java
@@ -58,7 +58,7 @@ public class PreparedStatementParameterTransformTest {
         });
         InterceptorHolder interceptorHolder = new InterceptorHolder(queryListener, queryTransformer, paramTransformer);
 
-        return new JdkJdbcProxyFactory().createConnection(rawDatasource.getConnection(), interceptorHolder);
+        return new JdkJdbcProxyFactory().createConnection(rawDatasource.getConnection(), interceptorHolder, "myDS");
     }
 
     @Test